Migrate App Infrastructure to Google App Engine
Considering the simple functionality required for
the service, and Google's generous free tier for
Google App Engine, I've rewritten the server to
run on GAE. I've also done so because a PaaS
offers substantially more stability with a
laissez-faire style management, compared to on an
IaaS as previous which often required some manual
intervention that I couldn't aptly give it.

In the event that the app breaks Google's
free tier quota, prices should be pennies (or
potentially fractions of a penny) per month.

Travis CI is being used as a continuous deployment
tool to push the master to production at
https://prtstat.us.

More open data retrieval features to come.
And maybe a wiki post or something with other
outlook for how other using other open source
metrics and information can better analyze events
and the system's overall effeciency.